chore: atom 및 selector 파일 분리
