+++ src/recoil/cartItems/atoms.ts
@@ -1,0 +1,16 @@
+import { atom, atomFamily } from 'recoil';
+
+import LocalStorage, { CART_ITEM } from '@/Storage';
+import { CartItemProps } from '@/types/cartItem';
+
+export const cartItemsState = atom<CartItemProps[]>({
+  key: 'cartItemsState',
+  default: [],
+});
+
+export const checkedItemsState = atomFamily<boolean, number>({
+  key: 'checkedItemsState',
+  default: (cartId) => {
+    return LocalStorage.getData(CART_ITEM, cartId) ?? true;
+  },
+});
