+++ src/types/cartItem.ts
@@ -1,0 +1,11 @@
+export interface CartItemProps {
+  id: number;
+  quantity: number;
+  product: {
+    id: number;
+    name: string;
+    price: number;
+    imageUrl: string;
+    category: string;
+  };
+}
